Moved NetworkRegistry.checksum() to HederaUtils.hip15checksum()
